Move cumulus/services/cip to services/cip
