+++ collections-online-cumulus/indexing/initialize/cip-client.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const cip = require('../../services/cip');
+const cip = require('../../../services/cip');
 const config = require('../../../collections-online/lib/config');
 
 /**
