+++ collections-online-cumulus/indexing/processing/result.js
@@ -5,7 +5,7 @@
  */
 
 const _ = require('lodash');
-const cip = require('../../services/cip');
+const cip = require('../../../services/cip');
 const config = require('../../../collections-online/lib/config');
 const es = require('../../../collections-online/lib/services/elasticsearch');
 const Q = require('q');
