+++ collections-online-cumulus/test/cip-binary-to-object.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var cip = require('../../services/cip');
+var cip = require('../../../services/cip');
 var fs = require('fs');
 var path = require('path');
 var assert = require('assert');
